# ZJIT: Add HIR for calling Cfunc with frame