--- a/zjit/src/hir.rs
+++ b/zjit/src/hir.rs
@@ -644,9 +644,12 @@ pub enum Insn {
     IfTrue { val: InsnId, target: BranchEdge },
     IfFalse { val: InsnId, target: BranchEdge },
 
-    /// Call a C function
+    /// Call a C function without pushing a frame
     /// `name` is for printing purposes only
     CCall { cfun: *const u8, args: Vec<InsnId>, name: ID, return_type: Type, elidable: bool },
+
+    /// Call a C function that pushes a frame
+    CCallWithFrame {cfun: *const u8, args: Vec<InsnId>, cme: *const rb_callable_method_entry_t, name: ID, state: InsnId },
 
     /// Call a variadic C function with signature: func(int argc, VALUE *argv, VALUE recv)
     /// This handles frame setup, argv creation, and frame teardown all in one
@@ -1022,6 +1025,13 @@ impl<'a> std::fmt::Display for InsnPrinter<'a> {
             Insn::GetConstantPath { ic, .. } => { write!(f, "GetConstantPath {:p}", self.ptr_map.map_ptr(ic)) },
             Insn::CCall { cfun, args, name, return_type: _, elidable: _ } => {
                 write!(f, "CCall {}@{:p}", name.contents_lossy(), self.ptr_map.map_ptr(cfun))?;
+                for arg in args {
+                    write!(f, ", {arg}")?;
+                }
+                Ok(())
+            },
+            Insn::CCallWithFrame { cfun, args, name, .. } => {
+                write!(f, "CallCFunc {}@{:p}", name.contents_lossy(), self.ptr_map.map_ptr(cfun))?;
                 for arg in args {
                     write!(f, ", {arg}")?;
                 }
@@ -1546,6 +1556,7 @@ impl Function {
             &ObjectAlloc { val, state } => ObjectAlloc { val: find!(val), state },
             &ObjectAllocClass { class, state } => ObjectAllocClass { class, state: find!(state) },
             &CCall { cfun, ref args, name, return_type, elidable } => CCall { cfun, args: find_vec!(args), name, return_type, elidable },
+            &CCallWithFrame { cfun, ref args, cme, name, state } => CCallWithFrame { cfun, args: find_vec!(args), cme, name, state: find!(state) },
             &CCallVariadic { cfun, recv, ref args, cme, name, state } => CCallVariadic {
                 cfun, recv: find!(recv), args: find_vec!(args), cme, name, state
             },
@@ -1646,6 +1657,7 @@ impl Function {
             Insn::NewRangeFixnum { .. } => types::RangeExact,
             Insn::ObjectAlloc { .. } => types::HeapObject,
             Insn::ObjectAllocClass { class, .. } => Type::from_class(*class),
+            Insn::CCallWithFrame { .. } => types::BasicObject,
             Insn::CCall { return_type, .. } => *return_type,
             Insn::CCallVariadic { .. } => types::BasicObject,
             Insn::GuardType { val, guard_type, .. } => self.type_of(*val).intersection(*guard_type),
@@ -2296,35 +2308,37 @@ impl Function {
                         return Err(Some(method));
                     }
 
+                    let ci_flags = unsafe { vm_ci_flag(call_info) };
+
+                    if ci_flags & VM_CALL_ARGS_SIMPLE == 0 {
+                        return Err(Some(method));
+                    }
+
+                    gen_patch_points_for_optimized_ccall(fun, block, recv_class, method_id, method, state);
+                    if recv_class.instance_can_have_singleton_class() {
+                        fun.push_insn(block, Insn::PatchPoint { invariant: Invariant::NoSingletonClass { klass: recv_class }, state });
+                    }
+                    if let Some(profiled_type) = profiled_type {
+                        // Guard receiver class
+                        recv = fun.push_insn(block, Insn::GuardType { val: recv, guard_type: Type::from_profiled_type(profiled_type), state });
+                    }
+                    let cfun = unsafe { get_mct_func(cfunc) }.cast();
+                    let mut cfunc_args = vec![recv];
+                    cfunc_args.append(&mut args);
+
                     // Filter for a leaf and GC free function
                     use crate::cruby_methods::FnProperties;
-                    let Some(FnProperties { leaf: true, no_gc: true, return_type, elidable }) =
-                        ZJITState::get_method_annotations().get_cfunc_properties(method)
-                    else {
-                        fun.set_dynamic_send_reason(send_insn_id, SendWithoutBlockCfuncNotVariadic);
-                        return Err(Some(method));
-                    };
-
-                    let ci_flags = unsafe { vm_ci_flag(call_info) };
                     // Filter for simple call sites (i.e. no splats etc.)
-                    if ci_flags & VM_CALL_ARGS_SIMPLE != 0 {
-                        gen_patch_points_for_optimized_ccall(fun, block, recv_class, method_id, method, state);
-
-                        if recv_class.instance_can_have_singleton_class() {
-                            fun.push_insn(block, Insn::PatchPoint { invariant: Invariant::NoSingletonClass { klass: recv_class }, state });
-                        }
-                        if let Some(profiled_type) = profiled_type {
-                            // Guard receiver class
-                            recv = fun.push_insn(block, Insn::GuardType { val: recv, guard_type: Type::from_profiled_type(profiled_type), state });
-                        }
-
-                        let cfun = unsafe { get_mct_func(cfunc) }.cast();
-                        let mut cfunc_args = vec![recv];
-                        cfunc_args.append(&mut args);
+                    // Commit to the replacement. Put PatchPoint.
+                    if let Some(FnProperties { leaf: true, no_gc: true, return_type, elidable }) = ZJITState::get_method_annotations().get_cfunc_properties(method) {
                         let ccall = fun.push_insn(block, Insn::CCall { cfun, args: cfunc_args, name: method_id, return_type, elidable });
                         fun.make_equal_to(send_insn_id, ccall);
-                        return Ok(())
+                    } else {
+                        let ccall = fun.push_insn(block, Insn::CCallWithFrame { cfun, args: cfunc_args, cme: method, name: method_id, state });
+                        fun.make_equal_to(send_insn_id, ccall);
                     }
+
+                    return Ok(());
                 }
                 // Variadic method
                 -1 => {
@@ -2637,7 +2651,8 @@ impl Function {
                 worklist.extend(args);
                 worklist.push_back(state);
             }
-            &Insn::InvokeBuiltin { ref args, state, .. }
+            &Insn::CCallWithFrame { ref args, state, .. }
+            | &Insn::InvokeBuiltin { ref args, state, .. }
             | &Insn::InvokeBlock { ref args, state, .. } => {
                 worklist.extend(args);
                 worklist.push_back(state)
@@ -10864,8 +10879,10 @@ mod opt_tests {
           Jump bb2(v4)
         bb2(v6:BasicObject):
           v11:HashExact = NewHash
-          v13:BasicObject = SendWithoutBlock v11, :dup
-          v15:BasicObject = SendWithoutBlock v13, :freeze
+          PatchPoint MethodRedefined(Hash@0x1000, dup@0x1008, cme:0x1010)
+          PatchPoint NoSingletonClass(Hash@0x1000)
+          v24:BasicObject = CallCFunc dup@0x1038, v11
+          v15:BasicObject = SendWithoutBlock v24, :freeze
           CheckInterrupts
           Return v15
         ");
@@ -10955,8 +10972,10 @@ mod opt_tests {
           Jump bb2(v4)
         bb2(v6:BasicObject):
           v11:ArrayExact = NewArray
-          v13:BasicObject = SendWithoutBlock v11, :dup
-          v15:BasicObject = SendWithoutBlock v13, :freeze
+          PatchPoint MethodRedefined(Array@0x1000, dup@0x1008, cme:0x1010)
+          PatchPoint NoSingletonClass(Array@0x1000)
+          v24:BasicObject = CallCFunc dup@0x1038, v11
+          v15:BasicObject = SendWithoutBlock v24, :freeze
           CheckInterrupts
           Return v15
         ");
@@ -11047,8 +11066,10 @@ mod opt_tests {
         bb2(v6:BasicObject):
           v10:StringExact[VALUE(0x1000)] = Const Value(VALUE(0x1000))
           v12:StringExact = StringCopy v10
-          v14:BasicObject = SendWithoutBlock v12, :dup
-          v16:BasicObject = SendWithoutBlock v14, :freeze
+          PatchPoint MethodRedefined(String@0x1008, dup@0x1010, cme:0x1018)
+          PatchPoint NoSingletonClass(String@0x1008)
+          v25:BasicObject = CallCFunc dup@0x1040, v12
+          v16:BasicObject = SendWithoutBlock v25, :freeze
           CheckInterrupts
           Return v16
         ");
@@ -11140,8 +11161,10 @@ mod opt_tests {
         bb2(v6:BasicObject):
           v10:StringExact[VALUE(0x1000)] = Const Value(VALUE(0x1000))
           v12:StringExact = StringCopy v10
-          v14:BasicObject = SendWithoutBlock v12, :dup
-          v16:BasicObject = SendWithoutBlock v14, :-@
+          PatchPoint MethodRedefined(String@0x1008, dup@0x1010, cme:0x1018)
+          PatchPoint NoSingletonClass(String@0x1008)
+          v25:BasicObject = CallCFunc dup@0x1040, v12
+          v16:BasicObject = SendWithoutBlock v25, :-@
           CheckInterrupts
           Return v16
         ");
@@ -11279,8 +11302,11 @@ mod opt_tests {
         bb2(v8:BasicObject, v9:BasicObject):
           v13:StringExact[VALUE(0x1000)] = Const Value(VALUE(0x1000))
           v25:BasicObject = GuardTypeNot v9, String
-          v26:BasicObject = SendWithoutBlock v9, :to_s
-          v17:String = AnyToString v9, str: v26
+          PatchPoint MethodRedefined(Array@0x1008, to_s@0x1010, cme:0x1018)
+          PatchPoint NoSingletonClass(Array@0x1008)
+          v30:ArrayExact = GuardType v9, ArrayExact
+          v31:BasicObject = CallCFunc to_s@0x1040, v30
+          v17:String = AnyToString v9, str: v31
           v19:StringExact = StringConcat v13, v17
           CheckInterrupts
           Return v19