@st0012 2025-09-29
I think the test failure is unrelated to the change, just surfaced by it. See https://github.com/Shopify/ruby/issues/768