@st0012 2025-10-03
> Also, have you considered adding profile information to opt_aref? That will get you more cfuncs.

I'll do that in a separate PR